cleanup and restructure
