+++ example/read_multi.py
@@ -7,6 +7,7 @@ This was tested against a S7-319 CPU
 import ctypes
 
 import snap7
+import snap7.util.getters
 from snap7.common import check_error
 from snap7.types import S7DataItem, S7AreaDB, S7WLByte
 from snap7 import util
@@ -55,7 +56,7 @@ for di in data_items:
 
 result_values = []
 # function to cast bytes to match data_types[] above
-byte_to_value = [util.get_real, util.get_real, util.get_int]
+byte_to_value = [snap7.util.getters.get_real, snap7.util.getters.get_real, snap7.util.getters.get_int]
 
 # unpack and test the result of each read
 for i in range(0, len(data_items)):
