+++ snap7/types.py
@@ -4,8 +4,6 @@ Python equivalent for snap7 specific types.
 import ctypes
 from enum import Enum
 
-from .common import ADict
-
 S7Object = ctypes.c_void_p
 buffer_size = 65536
 buffer_type = ctypes.c_ubyte * buffer_size
@@ -30,7 +28,7 @@ BRecvTimeout = 13
 RecoveryTime = 14
 KeepAliveTime = 15
 
-param_types = ADict({
+param_types = {
     LocalPort: ctypes.c_uint16,
     RemotePort: ctypes.c_uint16,
     PingTimeout: ctypes.c_int32,
@@ -46,7 +44,7 @@ param_types = ADict({
     BRecvTimeout: ctypes.c_int32,
     RecoveryTime: ctypes.c_uint32,
     KeepAliveTime: ctypes.c_uint32,
-})
+}
 
 # mask types
 mkEvent = 0
@@ -71,15 +69,14 @@ S7AreaDB = 0x84
 S7AreaCT = 0x1C
 S7AreaTM = 0x1D
 
-
-areas = ADict({
-    'PE': 0x81,
-    'PA': 0x82,
-    'MK': 0x83,
-    'DB': 0x84,
-    'CT': 0x1C,
-    'TM': 0x1D,
-})
+areas = {
+    'PE': S7AreaPE,
+    'PA': S7AreaPA,
+    'MK': S7AreaMK,
+    'DB': S7AreaDB,
+    'CT': S7AreaCT,
+    'TM': S7AreaTM,
+}
 
 
 # Word Length
@@ -117,16 +114,16 @@ srvAreaCT = 3
 srvAreaTM = 4
 srvAreaDB = 5
 
-server_areas = ADict({
-    'PE': 0,
-    'PA': 1,
-    'MK': 2,
-    'CT': 3,
-    'TM': 4,
-    'DB': 5,
-})
+server_areas = {
+    'PE': srvAreaPE,
+    'PA': srvAreaPA,
+    'MK': srvAreaMK,
+    'CT': srvAreaCT,
+    'TM': srvAreaTM,
+    'DB': srvAreaDB,
+}
 
-wordlen_to_ctypes = ADict({
+wordlen_to_ctypes = {
     S7WLBit: ctypes.c_int16,
     S7WLByte: ctypes.c_int8,
     S7WLWord: ctypes.c_int16,
@@ -134,9 +131,9 @@ wordlen_to_ctypes = ADict({
     S7WLReal: ctypes.c_int32,
     S7WLCounter: ctypes.c_int16,
     S7WLTimer: ctypes.c_int16,
-})
+}
 
-block_types = ADict({
+block_types = {
     'OB': ctypes.c_int(0x38),
     'DB': ctypes.c_int(0x41),
     'SDB': ctypes.c_int(0x42),
@@ -144,7 +141,7 @@ block_types = ADict({
     'SFC': ctypes.c_int(0x44),
     'FB': ctypes.c_int(0x45),
     'SFB': ctypes.c_int(0x46),
-})
+}
 
 server_statuses = {
     0: 'SrvStopped',
@@ -244,7 +241,7 @@ class S7DataItem(ctypes.Structure):
     ]
 
     def __str__(self) -> str:
-        return f"<S7DataItem Area: {self.Area} WordLen: {self.WordLen} Result: {self.Result} "\
+        return f"<S7DataItem Area: {self.Area} WordLen: {self.WordLen} Result: {self.Result} " \
                f"DBNumber: {self.DBNumber} Start: {self.Start} Amount: {self.Amount} pData: {self.pData}>"
 
 
@@ -258,7 +255,7 @@ class S7CpuInfo(ctypes.Structure):
     ]
 
     def __str__(self):
-        return f"<S7CpuInfo ModuleTypeName: {self.ModuleTypeName} SerialNumber: {self.SerialNumber} "\
+        return f"<S7CpuInfo ModuleTypeName: {self.ModuleTypeName} SerialNumber: {self.SerialNumber} " \
                f"ASName: {self.ASName} Copyright: {self.Copyright} ModuleName: {self.ModuleName}>"
 
 
@@ -312,7 +309,7 @@ class S7CpInfo(ctypes.Structure):
     ]
 
     def __str__(self) -> str:
-        return f"<S7CpInfo MaxPduLength: {self.MaxPduLength} MaxConnections: {self.MaxConnections} "\
+        return f"<S7CpInfo MaxPduLength: {self.MaxPduLength} MaxConnections: {self.MaxConnections} " \
                f"MaxMpiRate: {self.MaxMpiRate} MaxBusRate: {self.MaxBusRate}>"
